♻️ move timer and getZoneJsOriginalValue to @datadog/js-core/util
Both files already used globalObject from js-core; moving them avoids
a circular path when transport modules (flushController,
sendWithRetryStrategy) migrate to js-core in upcoming commits.

The js-core timer drops the monitor() wrapping — callers are responsible
for error handling at their own layer.
browser-core/tools/timer.ts keeps its monitored wrappers but now imports
getZoneJsOriginalValue from js-core.
browser-core/tools/getZoneJsOriginalValue.ts becomes a thin re-export.
