+++ packages/js-core/src/entries/util.ts
@@ -10,6 +10,10 @@ export * from '../util/jsonStringify'
 export { objectValues } from '../util/polyfills'
 export { isServerError } from '../util/responseUtils'
 export { mockable, mockableReplacements } from '../util/mockable'
+export type { BrowserWindowWithZoneJs } from '../util/getZoneJsOriginalValue'
+export { getZoneJsOriginalValue } from '../util/getZoneJsOriginalValue'
+export type { TimeoutId } from '../util/timer'
+export { setTimeout, clearTimeout, setInterval, clearInterval } from '../util/timer'
 export type { Subscription } from '../util/observable'
 export { Observable, BufferedObservable, mergeObservables } from '../util/observable'
 export type {
